backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
auth0-python==4.7.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==43.0.3
Django==4.2.18
django-auth-oidc==0.6.0
django-cors-headers==4.6.0
django-sslserver==0.22
djangorestframework==3.15.2
ecdsa==0.19.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
openid-connect==0.5.0
propcache==0.2.1
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
